test(e2e_appium): refuse peer-less runs per worker
Without STATUS_BACKEND_IMAGE a selection made only of peer tests fails
every selected test at setup with one message. The decision is taken in
the collection hook and stored per item, not raised there: under xdist
the hook runs on a worker, and an exception there is INTERNALERROR. A
mixed selection drops the peer tests, logs the count, and runs the rest.

The selection helpers live in core/peer_containers.py so the guards can
import them.
